+++ libraries/ast/mapping.h
@@ -330,6 +330,7 @@
 /* C header files. */
 /* --------------- */
 #include <stddef.h>
+#include <stdint.h>
 
 /* Macros. */
 /* ======= */
@@ -508,7 +509,7 @@ DECLARE_GENERIC_ALL(LD,long double)
                          const Xtype [], const Xtype [], int, const double [], \
                          int, double, int, Xtype, int, const int [], \
                          const int [], const int [], const int [], Xtype [], \
-                         Xtype [], double [], size_t *, int * );
+                         Xtype [], double [], int64_t *, int * );
 
 DECLARE_GENERIC_DFI(D,double)
 DECLARE_GENERIC_DFI(F,float)
@@ -607,7 +608,7 @@ PROTO_GENERIC_ALL(LD,long double)
                          const Xtype [], const Xtype [], int, const double [], \
                          int, double, int, Xtype, int, const int [], \
                          const int [], const int [], const int [], Xtype [], \
-                         Xtype [], double [], size_t *, int * );
+                         Xtype [], double [], int64_t *, int * );
 
 PROTO_GENERIC_DFI(D,double)
 PROTO_GENERIC_DFI(F,float)
